requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
nltk==3.8.1
pandas==2.1.4
python-dotenv==1.0.0
transformers==4.40.0
torch==2.1.0
sentencepiece==0.1.99
accelerate==0.25.0
bitsandbytes==0.41.3 